cats: use separate files for batch insert functions
