+++ core/src/cats/postgresql.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2003-2011 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2017 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -970,210 +970,6 @@ bool BareosDbPostgresql::SqlFieldIsNumeric(int field_type)
     default:
       return false;
   }
-}
-
-/**
- * Escape strings so that PostgreSQL is happy on COPY
- *
- *   NOTE! len is the length of the old string. Your new
- *         string must be long enough (max 2*old+1) to hold
- *         the escaped output.
- */
-static char* pgsql_copy_escape(char* dest, char* src, size_t len)
-{
-  /* we have to escape \t, \n, \r, \ */
-  char c = '\0';
-
-  while (len > 0 && *src) {
-    switch (*src) {
-      case '\n':
-        c = 'n';
-        break;
-      case '\\':
-        c = '\\';
-        break;
-      case '\t':
-        c = 't';
-        break;
-      case '\r':
-        c = 'r';
-        break;
-      default:
-        c = '\0';
-    }
-
-    if (c) {
-      *dest = '\\';
-      dest++;
-      *dest = c;
-    } else {
-      *dest = *src;
-    }
-
-    len--;
-    src++;
-    dest++;
-  }
-
-  *dest = '\0';
-  return dest;
-}
-
-bool BareosDbPostgresql::SqlBatchStartFileTable(JobControlRecord* jcr)
-{
-  const char* query = "COPY batch FROM STDIN";
-
-  Dmsg0(500, "SqlBatchStartFileTable started\n");
-
-  if (!SqlQueryWithoutHandler("CREATE TEMPORARY TABLE batch ("
-                              "FileIndex int,"
-                              "JobId int,"
-                              "Path varchar,"
-                              "Name varchar,"
-                              "LStat varchar,"
-                              "Md5 varchar,"
-                              "DeltaSeq smallint,"
-                              "Fhinfo NUMERIC(20),"
-                              "Fhnode NUMERIC(20))")) {
-    Dmsg0(500, "SqlBatchStartFileTable failed\n");
-    return false;
-  }
-
-  /*
-   * We are starting a new query.  reset everything.
-   */
-  num_rows_ = -1;
-  row_number_ = -1;
-  field_number_ = -1;
-
-  SqlFreeResult();
-
-  for (int i = 0; i < 10; i++) {
-    result_ = PQexec(db_handle_, query);
-    if (result_) { break; }
-    Bmicrosleep(5, 0);
-  }
-  if (!result_) {
-    Dmsg1(50, "Query failed: %s\n", query);
-    goto bail_out;
-  }
-
-  status_ = PQresultStatus(result_);
-  if (status_ == PGRES_COPY_IN) {
-    /*
-     * How many fields in the set?
-     */
-    num_fields_ = (int)PQnfields(result_);
-    num_rows_ = 0;
-    status_ = 1;
-  } else {
-    Dmsg1(50, "Result status failed: %s\n", query);
-    goto bail_out;
-  }
-
-  Dmsg0(500, "SqlBatchStartFileTable finishing\n");
-
-  return true;
-
-bail_out:
-  Mmsg1(errmsg, _("error starting batch mode: %s"), PQerrorMessage(db_handle_));
-  status_ = 0;
-  PQclear(result_);
-  result_ = NULL;
-  return false;
-}
-
-/**
- * Set error to something to abort operation
- */
-bool BareosDbPostgresql::SqlBatchEndFileTable(JobControlRecord* jcr,
-                                              const char* error)
-{
-  int res;
-  int count = 30;
-  PGresult* pg_result;
-
-  Dmsg0(500, "SqlBatchEndFileTable started\n");
-
-  do {
-    res = PQputCopyEnd(db_handle_, error);
-  } while (res == 0 && --count > 0);
-
-  if (res == 1) {
-    Dmsg0(500, "ok\n");
-    status_ = 1;
-  }
-
-  if (res <= 0) {
-    Dmsg0(500, "we failed\n");
-    status_ = 0;
-    Mmsg1(errmsg, _("error ending batch mode: %s"), PQerrorMessage(db_handle_));
-    Dmsg1(500, "failure %s\n", errmsg);
-  }
-
-  /*
-   * Check command status and return to normal libpq state
-   */
-  pg_result = PQgetResult(db_handle_);
-  if (PQresultStatus(pg_result) != PGRES_COMMAND_OK) {
-    Mmsg1(errmsg, _("error ending batch mode: %s"), PQerrorMessage(db_handle_));
-    status_ = 0;
-  }
-
-  PQclear(pg_result);
-
-  Dmsg0(500, "SqlBatchEndFileTable finishing\n");
-
-  return true;
-}
-
-bool BareosDbPostgresql::SqlBatchInsertFileTable(JobControlRecord* jcr,
-                                                 AttributesDbRecord* ar)
-{
-  int res;
-  int count = 30;
-  size_t len;
-  const char* digest;
-  char ed1[50], ed2[50], ed3[50];
-
-  esc_name = CheckPoolMemorySize(esc_name, fnl * 2 + 1);
-  pgsql_copy_escape(esc_name, fname, fnl);
-
-  esc_path = CheckPoolMemorySize(esc_path, pnl * 2 + 1);
-  pgsql_copy_escape(esc_path, path, pnl);
-
-  if (ar->Digest == NULL || ar->Digest[0] == 0) {
-    digest = "0";
-  } else {
-    digest = ar->Digest;
-  }
-
-  len = Mmsg(cmd, "%u\t%s\t%s\t%s\t%s\t%s\t%u\t%s\t%s\n", ar->FileIndex,
-             edit_int64(ar->JobId, ed1), esc_path, esc_name, ar->attr, digest,
-             ar->DeltaSeq, edit_uint64(ar->Fhinfo, ed2),
-             edit_uint64(ar->Fhnode, ed3));
-
-  do {
-    res = PQputCopyData(db_handle_, cmd, len);
-  } while (res == 0 && --count > 0);
-
-  if (res == 1) {
-    Dmsg0(500, "ok\n");
-    changes++;
-    status_ = 1;
-  }
-
-  if (res <= 0) {
-    Dmsg0(500, "we failed\n");
-    status_ = 0;
-    Mmsg1(errmsg, _("error copying in batch mode: %s"),
-          PQerrorMessage(db_handle_));
-    Dmsg1(500, "failure %s\n", errmsg);
-  }
-
-  Dmsg0(500, "SqlBatchInsertFileTable finishing\n");
-
-  return true;
 }
 
 /**
